python-tools/requirements.txt
# Verity Systems - Python Dependencies v5.0
# Production-Ready AI-Powered Fact-Checking Platform
# 100+ AI Models | Redis Caching | Circuit Breakers | Extended Data Sources

# ===========================================
# CORE AI/ML PROVIDERS
# ===========================================

# Anthropic Claude API (GitHub Education: $25/month credit)
anthropic==0.40.0

# Hugging Face Transformers and Hub (Free tier + GitHub Education credits)
transformers==4.36.0
huggingface-hub==0.20.0

# OpenAI-compatible APIs (Groq, Together, OpenRouter, CometAPI, etc.)
# CometAPI: 500+ models via OpenAI-compatible API
openai==1.12.0

# Google AI / Gemini
google-generativeai==0.3.2

# Mistral AI
mistralai==0.1.8

# Cohere
cohere==4.47

# Microsoft Azure AI Foundry
azure-ai-projects>=1.0.0
azure-identity>=1.25.0

# ===========================================
# UNIFIED LLM GATEWAY (NEW v5.0)
# ===========================================

# LiteLLM - 100+ LLM providers via unified interface
litellm>=1.80.0

# Ollama - Local LLM support
ollama>=0.3.0

# Groq - Ultra-fast inference (free tier)
groq>=0.4.0

# ===========================================
# ENHANCED SEARCH & DATA PROVIDERS
# ===========================================

# Tavily Search (AI-optimized)
tavily-python==0.3.0

# Exa (Neural Search)
exa-py==1.0.9

# SerpAPI
google-search-results==2.4.2

# ===========================================
# WEB FRAMEWORK & API
# ===========================================

# FastAPI - Modern async web framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
starlette==0.35.0

# HTTP client for async API calls
aiohttp==3.9.1
httpx==0.26.0

# Request validation
pydantic==2.5.3
email-validator==2.1.0

# ===========================================
# SECURITY & ENCRYPTION
# ===========================================

# Cryptography library for AES-256, PBKDF2
cryptography==41.0.7

# JWT token handling
PyJWT==2.8.0

# Password hashing (alternative to cryptography for passwords)
bcrypt==4.1.2

# Secrets management
python-dotenv==1.0.0

# ===========================================
# DATA PROCESSING
# ===========================================

# HTTP requests
requests==2.31.0

# JSON processing (faster than stdlib)
orjson==3.9.10

# Date/time handling
python-dateutil==2.8.2

# Progress bars
tqdm==4.66.4

# ===========================================
# NLP & TEXT PROCESSING
# ===========================================

# Language detection
langdetect==1.0.9

# Text cleaning
beautifulsoup4==4.12.2
lxml==5.1.0

# ===========================================
# CACHING & PERFORMANCE
# ===========================================

# In-memory caching
cachetools==5.3.2

# Redis client (for production distributed caching)
redis==5.0.1

# ===========================================
# MONITORING & LOGGING
# ===========================================

# Structured logging
structlog==24.1.0

# Sentry error tracking (GitHub Education: free tier)
sentry-sdk==1.39.1

# Honeybadger error tracking (Pro trial from GitHub Education)
honeybadger==0.13.0

# New Relic APM monitoring (Free tier from GitHub Education)
newrelic==9.6.0

# ConfigCat feature flags (Free tier for gradual rollouts)
configcat-client==9.0.3

# Comet ML - ML Experiment Tracking (GitHub Education: free tier)
comet-ml==3.35.0

# ===========================================
# DATABASE (Optional - for persistent storage)
# ===========================================

# SQLAlchemy for database ORM
sqlalchemy==2.0.25

# PostgreSQL adapter
psycopg2-binary==2.9.9

# MongoDB driver (GitHub Education: MongoDB Atlas free tier)
pymongo==4.6.1

# Payment Processing
stripe==5.13.0

# ===========================================
# TESTING
# ===========================================

pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# ===========================================
# TYPE CHECKING
# ===========================================

typing-extensions==4.9.0

# ===========================================
# UTILITIES
# ===========================================

# Rate limiting
slowapi==0.1.9

# Retry logic
tenacity==8.2.3

# Environment variable parsing
environs==10.3.0

# ===========================================
# AWS SERVICES (LocalStack for local dev)
# ===========================================

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# ===========================================
# GEOGRAPHIC & VISUALIZATION
# ===========================================

# Carto integration (Location Intelligence)
# Uses REST API - no specific package needed, using aiohttp

# IP Geolocation
# Uses ipinfo.io API via aiohttp
# ===========================================
# RESILIENCE & OBSERVABILITY (NEW v5.0)
# ===========================================

# Prometheus metrics client
prometheus-client==0.19.0

# Circuit breaker pattern
pybreaker==1.0.1

# Distributed tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-aiohttp-client==0.43b0

# ===========================================
# ASYNC UTILITIES (NEW v5.0)
# ===========================================

# Async Redis
aioredis==2.0.1

# Async utilities
aiocache==0.12.2
aiofiles==23.2.1

# ===========================================
# API INTEGRATIONS (ENHANCED v5.0)
# ===========================================

# Together AI
together>=0.2.0

# DeepSeek API (OpenAI compatible, uses openai package)

# OpenRouter (OpenAI compatible, uses openai package)